# specs/001-okr-crud-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: OKR Management API
  description: RESTful API for managing Objectives and Key Results (OKRs)
  version: 1.0.0
  contact:
    name: Okra API Team
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /time-periods:
    get:
      tags: [time-periods]
      operationId: listTimePeriods
      summary: List time periods
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TimePeriodStatus'
      responses:
        '200':
          description: List of time periods
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TimePeriodResponse'
    
    post:
      tags: [time-periods]
      operationId: createTimePeriod
      summary: Create a new time period
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimePeriodRequest'
      responses:
        '201':
          description: Time period created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimePeriodResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /time-periods/{timePeriodId}:
    parameters:
      - $ref: '#/components/parameters/TimePeriodIdParam'
    
    get:
      tags: [time-periods]
      operationId: getTimePeriod
      summary: Get time period by ID
      responses:
        '200':
          description: Time period details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimePeriodResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    
    patch:
      tags: [time-periods]
      operationId: updateTimePeriod
      summary: Update a time period
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTimePeriodRequest'
      responses:
        '200':
          description: Time period updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimePeriodResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
  /time-periods/{timePeriodId}/archive:
    parameters:
      - $ref: '#/components/parameters/TimePeriodIdParam'
    
    post:
      tags: [time-periods]
      operationId: archiveTimePeriod
      summary: Archive a time period
      responses:
        '200':
          description: Time period archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimePeriodResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============ REPORTS ============
...
components:
  parameters:
...
    TimePeriodIdParam:
      name: timePeriodId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    TimePeriodStatus:
      type: string
      enum: [active, archived]

    # ===== REQUEST SCHEMAS =====
...
    CreateTimePeriodRequest:
      type: object
      required: [name, startDate, endDate]
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    UpdateTimePeriodRequest:
      type: object
      required: [version]
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        version:
          type: integer

    # ===== RESPONSE SCHEMAS =====
...
    TimePeriodResponse:
      type: object
      required: [id, name, startDate, endDate, status, createdAt, updatedAt, version]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/TimePeriodStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              additionalProperties: true
              description: Additional error context
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: Request validation failed
              details:
                violations:
                  - field: endDate
                    message: End date must be after start date
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []